backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Celery
redis==5.0.1
celery==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Network Scanning
python-nmap==0.7.1
scapy==2.5.0

# Payment Processing
stripe==7.8.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
jinja2==3.1.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# File Processing
python-magic==0.4.27
pillow==10.1.0

# Async Support
asyncpg==0.29.0
aiofiles==23.2.1